reverts ensure pubsub message exchange to internal
