+++ network/internal/p2pfixtures/fixtures.go
@@ -5,7 +5,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"github.com/onflow/flow-go/network/p2p/test"
+	"github.com/onflow/flow-go/network/message"
 	"net"
 	"testing"
 	"time"
@@ -259,7 +259,7 @@ func EnsureNoPubsubMessageExchange(t *testing.T, ctx context.Context, from []p2p
 		msg, _ := messageFactory()
 		channel, ok := channels.ChannelFromTopic(topic)
 		require.True(t, ok)
-		data := p2ptest.MustEncodeEvent(t, msg, channel)
+		data := MustEncodeEvent(t, msg, channel)
 
 		// ensure the message is NOT received by any of the nodes.
 		require.NoError(t, node.Publish(ctx, topic, data))
@@ -367,5 +367,82 @@ func LongStringMessageFactoryFixture(t *testing.T) func() string {
 		msg := "this is an intentionally long MESSAGE to be bigger than buffer size of most of stream compressors"
 		require.Greater(t, len(msg), 10, "we must stress test with longer than 10 bytes messages")
 		return fmt.Sprintf("%s %d \n", msg, time.Now().UnixNano()) // add timestamp to make sure we don't send the same message twice
+	}
+}
+
+// EnsurePubsubMessageExchange ensures that the given nodes exchange the given message on the given channel through pubsub.
+func EnsurePubsubMessageExchange(t *testing.T, ctx context.Context, nodes []p2p.LibP2PNode, messageFactory func() (interface{}, channels.Topic)) {
+	_, topic := messageFactory()
+
+	subs := make([]p2p.Subscription, len(nodes))
+	slashingViolationsConsumer := unittest.NetworkSlashingViolationsConsumer(unittest.Logger(), metrics.NewNoopCollector())
+	for i, node := range nodes {
+		ps, err := node.Subscribe(
+			topic,
+			validator.TopicValidator(
+				unittest.Logger(),
+				unittest.NetworkCodec(),
+				slashingViolationsConsumer,
+				unittest.AllowAllPeerFilter()))
+		require.NoError(t, err)
+		subs[i] = ps
+	}
+
+	// let subscriptions propagate
+	time.Sleep(1 * time.Second)
+
+	channel, ok := channels.ChannelFromTopic(topic)
+	require.True(t, ok)
+
+	for _, node := range nodes {
+		// creates a unique message to be published by the node
+		msg, _ := messageFactory()
+		data := MustEncodeEvent(t, msg, channel)
+		require.NoError(t, node.Publish(ctx, topic, data))
+
+		// wait for the message to be received by all nodes
+		ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
+		SubsMustReceiveMessage(t, ctx, data, subs)
+		cancel()
+	}
+}
+
+// MustEncodeEvent encodes and returns the given event and fails the test if it faces any issue while encoding.
+func MustEncodeEvent(t *testing.T, v interface{}, channel channels.Channel) []byte {
+	bz, err := unittest.NetworkCodec().Encode(v)
+	require.NoError(t, err)
+
+	msg := message.Message{
+		ChannelID: channel.String(),
+		Payload:   bz,
+	}
+	data, err := msg.Marshal()
+	require.NoError(t, err)
+
+	return data
+}
+
+// SubMustReceiveMessage checks that the subscription have received the given message within the given timeout by the context.
+func SubMustReceiveMessage(t *testing.T, ctx context.Context, expectedMessage []byte, sub p2p.Subscription) {
+	received := make(chan struct{})
+	go func() {
+		msg, err := sub.Next(ctx)
+		require.NoError(t, err)
+		require.Equal(t, expectedMessage, msg.Data)
+		close(received)
+	}()
+
+	select {
+	case <-received:
+		return
+	case <-ctx.Done():
+		require.Fail(t, "timeout on receiving expected pubsub message")
+	}
+}
+
+// SubsMustReceiveMessage checks that all subscriptions receive the given message within the given timeout by the context.
+func SubsMustReceiveMessage(t *testing.T, ctx context.Context, expectedMessage []byte, subs []p2p.Subscription) {
+	for _, sub := range subs {
+		SubMustReceiveMessage(t, ctx, expectedMessage, sub)
 	}
 }
