+++ network/p2p/scoring/app_score_test.go
@@ -91,14 +91,14 @@ func TestFullGossipSubConnectivity(t *testing.T) {
 	// checks end-to-end message delivery works
 	// each node sends a distinct message to all and checks that all nodes receive it.
 	for _, node := range nodes {
-		proposalMsg := p2ptest.MustEncodeEvent(t, unittest.ProposalFixture(), channels.PushBlocks)
+		proposalMsg := p2pfixtures.MustEncodeEvent(t, unittest.ProposalFixture(), channels.PushBlocks)
 		require.NoError(t, node.Publish(ctx, blockTopic, proposalMsg))
 
 		// checks that the message is received by all nodes.
 		ctx1s, cancel1s := context.WithTimeout(ctx, 5*time.Second)
-		p2ptest.SubsMustReceiveMessage(t, ctx1s, proposalMsg, groupOneSubs)
-		p2ptest.SubsMustReceiveMessage(t, ctx1s, proposalMsg, accessNodeSubs)
-		p2ptest.SubsMustReceiveMessage(t, ctx1s, proposalMsg, groupTwoSubs)
+		p2pfixtures.SubsMustReceiveMessage(t, ctx1s, proposalMsg, groupOneSubs)
+		p2pfixtures.SubsMustReceiveMessage(t, ctx1s, proposalMsg, accessNodeSubs)
+		p2pfixtures.SubsMustReceiveMessage(t, ctx1s, proposalMsg, groupTwoSubs)
 
 		cancel1s()
 	}
@@ -201,7 +201,7 @@ func testGossipSubMessageDeliveryUnderNetworkPartition(t *testing.T, honestPeerS
 	// let nodes reside on a full topology, hence no partition is caused by the topology.
 	p2ptest.LetNodesDiscoverEachOther(t, ctx, allNodes, allIds)
 
-	proposalMsg := p2ptest.MustEncodeEvent(t, unittest.ProposalFixture(), channels.PushBlocks)
+	proposalMsg := p2pfixtures.MustEncodeEvent(t, unittest.ProposalFixture(), channels.PushBlocks)
 	require.NoError(t, con1Node.Publish(ctx, blockTopic, proposalMsg))
 
 	// we check that whether within a one-second window the message is received by the other honest consensus node.
